requirements.txt
numpy
python-dotenv
discord.py
beautifulsoup4
pandas
tabulate